+++ src/generators/Generators.jl
@@ -10,7 +10,6 @@ using LinearAlgebra
 using ..Models
 using ..Objectives
 using Statistics
-using Parameters
 using DecisionTree
 using DataFrames
 using MLJBase